tarmac/new_pattern24/app-shared-protobuffs/python/pyproject.toml
[tool.poetry]
name = "cn-protos"

version = "2.1.0"
description = "Python library containing the Clinician Nexus Shared Proto3s along with some utility functions."
authors = ["Data Platform <data-platform@cliniciannexus.com>"]
readme = "README.md"

# override gitignore
include = ["src/buf/**/*.py*"]
packages = [
    { include = "cn_proto_utils", from = "src"},
    { include = "buf", from = "src"} # file descriptor API
]

[tool.poetry.dependencies]
python = "^3.10"
protobuf = "^4.24.4"
types-protobuf = "^4.24.0.3"
pytest = "^7.4.3"
grpcio = "^1.59.2"
pyspark = {version = ">=3.5.0", optional = true}
cn-shared-app-shared-protobuffs-protocolbuffers-python = "^25.2.0.1.20240202025918+7cac801cf13e"


[tool.poetry.extras]
pyspark = ["pyspark"]


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "buf"
url = "https://clinician-nexus.buf.dev/gen/python"
